REFACT :: Jwt.kt에서 JwtType 분리
